Enables moduli 62 bits parameters
If 62 bits is chosen for moduli P, it will select primes which are close to 2^62 but of 61 bits. To be used for moduli P when 60 bit primes are used for moduli Q. This enable a better noise reduction during the key-switch, while still keeping all the moduli under 62 bits (ensuring no overflow during the NTT).